docs: add original concept document to docs dir
The concept document pdf is (at least supposed to be) stored via Git LFS
in order to not bloat the repository.

The document serves as a useful reference to the (medium-term) goals
for this project but since it is a personal historical document for
this project, I've put it under a very restrictive CC-license to allow
for redistribution but prevent other shenanigans.