feat: implement mechanism for enabling/disabling mods
You can now enable or disable a mod by using the command
modfs.py enable modid
or
modfs.py disable modid

In order to store this information, a new configuration system for
mod-specific information has been created.
This data will be stored in the mods/ directory in a JSON file
whose name follows the schema <modid>.json

Additional data fields for information such as author name,
version selection or a prettier mod name are considered valid settings
but currently not recognized or used by the software - this may be
added in the future.

Resolves https://github.com/jotoho/modfs/issues/8
Signed-off-by: Jonas Tobias Hopusch <git@jotoho.de>