feat: implement mechanism for manual mod version selection
There are many situations where one may want to use a different
version than the latest. for example, for the purposes of debugging
a mod loadout or if certain versions are incompatible with each other.

In this new version it is now possible to set a specific version of a
mod to use via the
modfs.py useversion <modid> <version_date>/<subversion>
command.

Alternatively, executing
modfs.py useversion <modid> latest
will return that mod to the default selection algorithm.

Signed-off-by: Jonas Tobias Hopusch <git@jotoho.de>