fix: prevent crash when importing a mod without prior installed versions
An oversight in the programming of a function used to make sure
we do not exceed the number of allowed subversions caused it to
wrongly request version information on a mod that does not exist
and caused the app to crash.

Skipping this step when the mod is not yet installed resolves this
issue.

Signed-off-by: Jonas Tobias Hopusch <git@jotoho.de>