feat: add ability to track last update check as mod metadata
This commit introduces a new value to the mod metadata file format
"last_update_check" which can contain the date of the last time the mod
was updated or manually marked as up-to-date using the new subcommand
"markuptodate".

In order to make this new information fields useful, the commit also
adds the subcommands "list updatecheck", which prints a formatted
list of update dates sorted chronologically and their respective mods.
The last updated date of a specific mod has also been added to the
output of the "mod <mod> info" subcommand.
If no last updated date is saved, modfs falls back to the date of the
latest currently installed mod - please be aware that this fallback
may be impacted by the deletion of specific versions.

If neither the metadata field, nor any version of the mod exists,
there's some special case handling attached to gracefully deal with the
issue.

Signed-off-by: Jonas Tobias Hopusch <git@jotoho.de>